增加管理员编辑，删除等功能

+++ infrastructure/src/main/java/edu/jxau/cjn/infrastructure/repositories/UserRepository.java
@@ -4,7 +4,7 @@ import edu.jxau.cjn.infrastructure.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.repository.PagingAndSortingRepository;
 
-public interface UserRepository extends JpaRepository<User, Integer>, PagingAndSortingRepository<User, Integer> {
+public interface UserRepository extends JpaRepository<User, Long>, PagingAndSortingRepository<User, Long> {
 
     //根据学号查询学生
     User findUserByNickName(String nickName);
